% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_autocorrelation_lag1_2nd_val.R
\name{f.autocorrelation.lag1.2nd.val}
\alias{f.autocorrelation.lag1.2nd.val}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.autocorrelation.lag1.2nd.val(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
