% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_autocorrelation_lag2_1st_val.R
\name{f.autocorrelation.lag2.1st.val}
\alias{f.autocorrelation.lag2.1st.val}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.autocorrelation.lag2.1st.val(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
