% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_autocorrelation_lag2_string.R
\name{f.autocorrelation.lag2.string}
\alias{f.autocorrelation.lag2.string}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.autocorrelation.lag2.string(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
