% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_timeseries.R
\name{f.autocorrelation}
\alias{f.autocorrelation}
\title{A Time-Series Feature Extraction Package}
\usage{
f.autocorrelation(x, lag)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
A Time-Series Feature Extraction Package
}
