% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_coeff_variation.R
\name{f.coeff.variation}
\alias{f.coeff.variation}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.coeff.variation(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
...
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
