% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_count_above_3sigma.R
\name{f.count.above.3sigma}
\alias{f.count.above.3sigma}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.count.above.3sigma(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
