% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_count_below_1sigma.R
\name{f.count.below.1sigma}
\alias{f.count.below.1sigma}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.count.below.1sigma(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
