% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_count_NA.R
\name{f.count.NA}
\alias{f.count.NA}
\alias{f.count_NA}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.count_NA(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
custom function to count NAs
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
