% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_count_negative.R
\name{f.count.negative}
\alias{f.count.negative}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.count.negative(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
