% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_count_peaks.R
\name{f.count.peaks}
\alias{f.count.peaks}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.count.peaks(x, m = 3)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
