% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_entropy.R
\name{f.entropy}
\alias{f.entropy}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.entropy(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns Shannon Entropy
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
