% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_first_max.R
\name{f.first.max}
\alias{f.first.max}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.first.max(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
