% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_location.R
\name{f.first.min}
\alias{f.first.min}
\title{A Time-Series Feature Extraction Package}
\usage{
f.first.min(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
A Time-Series Feature Extraction Package
}
