% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_IQR.R
\name{f.IQR}
\alias{f.IQR}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.IQR(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
calculate IQR of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
