% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_kurtosis.R
\name{f.kurtosis}
\alias{f.kurtosis}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.kurtosis(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
