% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_meta.R
\name{f.length}
\alias{f.length}
\title{A Time-Series Feature Extraction Package}
\usage{
f.length(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
calculate length of x

Returns the length of observations
}
\description{
A Time-Series Feature Extraction Package
}
