% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_max.R
\name{f.max}
\alias{f.max}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.max(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
calculate max of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
