% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mean.R
\name{f.mean}
\alias{f.mean}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.mean(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
calculate median of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
