% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_median_change.R
\name{f.median.change}
\alias{f.median.change}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.median.change(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns the median over the absolute differences between subsequent time series values which is
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
