% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_median.R
\name{f.median}
\alias{f.median}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.median(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
calculate median of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
