% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_min_change.R
\name{f.min.change}
\alias{f.min.change}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.min.change(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
Returns the min over the absolute differences between subsequent time series values which is
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
