% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_min.R
\name{f.min}
\alias{f.min}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.min(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
calculate min of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
