% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_change.R
\name{f.mssd.lag1}
\alias{f.mssd.lag1}
\title{A Time-Series Feature Extraction Package}
\usage{
f.mssd.lag1(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns the MSSD, lag1; Von Neuman et al. (1941) discussed the Mean Square of Successive Differences as a measure of variability that takes into account gradual shifts in mean
}
\description{
A Time-Series Feature Extraction Package
}
