% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mssd_lag2.R
\name{f.mssd.lag2}
\alias{f.mssd.lag2}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.mssd.lag2(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns the MSSD, lag2; Von Neuman et al. (1941) discussed the Mean Square of Successive Differences as a measure of variability that takes into account gradual shifts in mean
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
