% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_partial_autocorrelation_lag1_string.R
\name{f.partial.autocorrelation.lag1.string}
\alias{f.partial.autocorrelation.lag1.string}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.partial.autocorrelation.lag1.string(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
