% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_partial_autocorrelation_lag2_1st_val.R
\name{f.partial.autocorrelation.lag2.1st.val}
\alias{f.partial.autocorrelation.lag2.1st.val}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.partial.autocorrelation.lag2.1st.val(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
