% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_perc_missing.R
\name{f.perc.missing}
\alias{f.perc.missing}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.perc.missing(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
percentage missing data
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
