% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prop_missing.R
\name{f.prop.missing}
\alias{f.prop.missing}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.prop.missing(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
percentage missing data
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
