% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_quantile_20.R
\name{f.quantile_20}
\alias{f.quantile_20}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.quantile_20(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
