% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_quantile_60.R
\name{f.quantile_60}
\alias{f.quantile_60}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.quantile_60(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
