% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_quantile_75.R
\name{f.quantile_75}
\alias{f.quantile_75}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.quantile_75(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
