% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_range.R
\name{f.range}
\alias{f.range}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.range(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
Range of data
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
