% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ratio_unique_values.R
\name{f.ratio.unique.values}
\alias{f.ratio.unique.values}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.ratio.unique.values(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns the ratio of unique values, that are present in the time series more than once.
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
