% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sd.R
\name{f.sd}
\alias{f.sd}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.sd(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
calculate sd of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
