% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_skew.R
\name{f.skew}
\alias{f.skew}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.skew(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
