% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sum_change.R
\name{f.sum.change}
\alias{f.sum.change}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.sum.change(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns the sum over the absolute differences between subsequent time series values which is
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
