% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sum.R
\name{f.sum}
\alias{f.sum}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.sum(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
calculate sum of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
