% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_variance.R
\name{f.variance}
\alias{f.variance}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.variance(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
calculate variance of x
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
