% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_which_above_2sigma.R
\name{f.which.above.2sigma}
\alias{f.which.above.2sigma}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.which.above.2sigma(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
