% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_which_below_1sigma.R
\name{f.which.below.1sigma}
\alias{f.which.below.1sigma}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.which.below.1sigma(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
