% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_corr_heatmap.R
\name{feature_corr_heatmap}
\alias{feature_corr_heatmap}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
feature_corr_heatmap(cor.df, out.path = NA, palette = "viridis",
  theme = "minimal")
}
\arguments{
\item{out.path}{class: string; where to save correlation heatmap}

\item{palette}{class: string; options: "viridis", "temperature"}

\item{theme}{class: string; options: "minimal", "blank"}

\item{df}{class: data.frame}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
\examples{
feature_corr_heatmap(df, out.path)
}
\keyword{intra-individual}
\keyword{series,}
\keyword{time}
\keyword{variability}
