% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_correlations.R
\name{feature_correlations}
\alias{feature_correlations}
\title{A Time-Series Feature Extraction Package}
\usage{
feature_correlations(df, data.format = "wide", group_var, plot = T,
  view = F, r_threshold = 0.5)
}
\arguments{
\item{df}{class: data.frame}

\item{plot}{class: boolean; whether or not to plot feature correlations as heatmap}

\item{view}{class: boolean; whether or not to show data.frame of highly correlated variables above preset threshold (default = 0.5)}

\item{r_threshold}{class: numeric; correlation threshold to display above}
}
\description{
A Time-Series Feature Extraction Package
}
\examples{
feature_correlations(df, view=T, r_threshold=0.75)
COMING SOON: p-value support
}
\keyword{intra-individual}
\keyword{series,}
\keyword{time}
\keyword{variability}
