% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_to_df.R
\name{features_to_df}
\alias{features_to_df}
\title{A Time-Series Feature Extraction Package}
\usage{
features_to_df(df.list, group_var, data.format = "long", verbose = F)
}
\arguments{
\item{df.list}{class: list of dataframes; each dataframe result of different feature calculator}

\item{group_var}{class: vector of strings to group data by (ID, variable)}

\item{data.format}{class: string; 'wide' or 'long'; currently no support for 'wide'}

\item{verbose}{class: boolean; T = display status, debugging info}
}
\description{
A Time-Series Feature Extraction Package
}
\examples{
features_to_df(df.list, group_var='id', data.format='long')
features_to_df(df.list, group_var=c('id','session'), data.format='wide')
}
\keyword{intra-individual}
\keyword{series,}
\keyword{time}
\keyword{variability}
