% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_extraction_list.R
\name{generate_extraction_list}
\alias{generate_extraction_list}
\title{A Time-Series Feature Extraction Package}
\usage{
generate_extraction_list(feature_category)
}
\arguments{
\item{feature_category}{class: string; options: 'all', 'simple', 'count', 'sigma', change', 'meta', 'distribution' , 'location'}
}
\description{
A Time-Series Feature Extraction Package
}
\examples{
generate_extraction_list(feature_category="simple")
generate_extraction_list(feature_category="all")
}
