% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{A Time-Series Feature Extraction Package}
\usage{
plot_series(df, ids = c(1:2), id_var = "id", x = "day", y = "y")
}
\arguments{
\item{df}{class: data.frame}

\item{ids}{class: vector of user ids to match}

\item{id_var}{class: string; description: column name with value 'y' in data}

\item{x}{class: string; description: column name with value 'x' in data}

\item{y}{class: string; description: column name with value 'y' in data}
}
\description{
A Time-Series Feature Extraction Package
}
\examples{
plot_series(isahib.sim, ids=c(1:3), id_var="id", x="day", y="y")
}
