% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prob_acute_change_6.R
\name{prob.acute.change.6}
\alias{prob.acute.change.6}
\alias{f.prob.acute.change.6}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.prob.acute.change.6(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
Returns the probability of acute change (threshold = 0.60)
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
