% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prob_acute_change_7.R
\name{prob.acute.change.7}
\alias{prob.acute.change.7}
\alias{f.prob.acute.change.7}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.prob.acute.change.7(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
Returns the probability of acute change (threshold = 0.70)
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
