% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_autocorrelation.R
\name{f.autocorrelation}
\alias{f.autocorrelation}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.autocorrelation(x, lag)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
