% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_count.R
\name{f.count.negative}
\alias{f.count.negative}
\title{A Time-Series Feature Extraction Package}
\usage{
f.count.negative(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
A Time-Series Feature Extraction Package
}
