% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_first_min.R
\name{f.first.min}
\alias{f.first.min}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.first.min(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
