% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_distribution.R
\name{f.kurtosis}
\alias{f.kurtosis}
\title{A Time-Series Feature Extraction Package}
\usage{
f.kurtosis(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
A Time-Series Feature Extraction Package
}
