% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_length.R
\name{f.length}
\alias{f.length}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
f.length(x)
}
\arguments{
\item{x}{class: vector, data from dplyr}
}
\value{
calculate length of x

Returns the length of observations
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
