% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_binary.R
\name{f.sd.greater.range}
\alias{f.sd.greater.range}
\title{A Time-Series Feature Extraction Package}
\usage{
f.sd.greater.range(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
if sd is greater than range of data
}
\description{
A Time-Series Feature Extraction Package
}
