% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_sigma.R
\name{f.which.below.1sigma}
\alias{f.which.below.1sigma}
\title{A Time-Series Feature Extraction Package}
\usage{
f.which.below.1sigma(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\description{
A Time-Series Feature Extraction Package
}
