% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{`tsfeaturex: A Time-Series Feature Extraction Package}
\usage{
normalize(x)
}
\arguments{
\item{x}{class: vector, data from dplyr %>%}
}
\value{
transforms data to 0-1
}
\description{
`tsfeaturex: A Time-Series Feature Extraction Package
}
